requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap-icons==0.5.5
django-bootstrap3==15.0.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-tinymce==3.3.0
idna==2.10
pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4

python-crontab~=2.5.1